Move grid layout logic out of the file that imports @expo/ui
columnsForFontScale and inRows never touch @expo/ui, but living in
tile.tsx meant testing them required mocking the SwiftUI module --
asserting the mock's behaviour rather than the real logic. Moving both
into tile-layout.ts, alongside TILE_SPACING, lets their Jest test
import the real module directly, with nothing to fake.
